% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_initials_cluster.R
\name{get_initial}
\alias{get_initial}
\title{Obtain initial values for estimation procedure.}
\usage{
get_initial(dat, K)
}
\arguments{
\item{dat}{a data frame containing the data in the model}

\item{K}{number of latent classes}
}
\value{
A list containing the following components:
\tabular{ll}{
   \code{ini_alpha} \tab A matrix of initial alpha \cr
   \tab \cr
   \code{ini_beta} \tab A matrix of initial beta \cr
}
}
\description{
Obtain initial value for alpha by: 1. assign class membership to all subjects with Kmeans, then fit the multinomial regression to obtain alpha.
Obtain initial value for beta by fitting the multiplicative intensity model studies by Wang et al. (2001) using the reReg() function, stratified by the latent class membership assigned by Kmeans.
}
