% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mu_t.R
\name{mu_t}
\alias{mu_t}
\title{estimate mu(t)}
\usage{
mu_t(time_long, censor_long, t)
}
\arguments{
\item{time_long}{long format time - events (excluding censoring time)}

\item{censor_long}{long format censoring time (longest follow up time)}

\item{t}{time of interest}
}
\value{
estimated mu(t)
}
\description{
Estimate mu0 with the Nelson-Aalen type estimator under the assumed multiplicative intensity modeling of recurrent events
}
