% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_p_D.R
\name{p_D}
\alias{p_D}
\title{Estimate P(D|xi, Z, C)}
\usage{
p_D(d, beta, Z, mu_censor, gamma = 0)
}
\arguments{
\item{d}{a vector of observed recurrent events for subjects of interest}

\item{beta}{class specific parameters for recurrent model}

\item{Z}{a vector of time-independent corvariates}

\item{mu_censor}{a vector of estimated mu(C), where C is a vector of censoring time}

\item{individual}{frailty. 0 represents the frailty equals 1 and k reprsents the frailty follows gamma(k,k)}
}
\value{
a vector of estimated P(Di_di | xi = k, Zi, Ci)
}
\description{
Estimate P(Di = di | xi = k, Zi, Ci)
}
