% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_p_xi.R
\name{p_xi}
\alias{p_xi}
\title{Estimate P(xi | Z, C)}
\usage{
p_xi(alpha, Z)
}
\arguments{
\item{alpha}{regression coefficient for multinomial logistic regression model (xi)}

\item{Z}{a vector of time-independent corvariates}
}
\value{
a vector of estimated P(xi = k | Zi, Ci)
}
\description{
estimate P(xi = k | Zi, Ci)
}
