% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{predict.SLCARE}
\alias{predict.SLCARE}
\title{Predict Results for SLCARE}
\usage{
\method{predict}{SLCARE}(object, integer = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{SLCARE}.}

\item{integer}{logicals. If \code{FALSE} (the default value), the function returns to float numbers.}

\item{...}{other arguments.}
}
\description{
Calculate the posterior predicted number of recurrent events.
}
\seealso{
\code{\link[=SLCARE]{SLCARE()}}
}
