% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessingdata.R
\name{PreprocessData}
\alias{PreprocessData}
\title{data pre-processing}
\usage{
PreprocessData(dat = dat)
}
\arguments{
\item{dat}{a data frame containing the data in the model}
}
\value{
A list containing the following components:
\tabular{ll}{
   \code{id_wide} \tab A vector of subjects ID in wide format \cr
   \tab \cr
   \code{d} \tab A vector of observed recurrent events for subjects of interest \cr
   \tab \cr
   \code{Z} \tab A vector of time-independent corvariates in wide format \cr
   \tab \cr
   \code{censor_wide} \tab A vector of censoring time (longest follow up time) in wide format \cr
   \tab \cr
   \code{id_long} \tab A vector of subjects ID in long format \cr
   \tab \cr
   \code{time_long} \tab A vector of long format time for events (excluding censoring time) \cr
   \tab \cr
   \code{censor_long} \tab A vector of censoring time (longest follow up time) in long format \cr
   \tab \cr
   \code{Z_long} \tab A vector of time-independent corvariates in long format \cr
}
}
\description{
pre-process data into long and wide formats that fit functions in this R package
}
