% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.SLCARE}
\alias{print.SLCARE}
\title{Print Results for SLCARE}
\usage{
\method{print}{SLCARE}(
  x,
  type = c("Est", "SE", "Init", "ConvergeLoss", "ClassProb", "PostPredict", "Entropy",
    "PValue"),
  ...
)
}
\arguments{
\item{x}{an object of class \code{SLCARE}.}

\item{type}{the type of the output.}

\item{...}{other arguments.}
}
\description{
Print point estimates (Est), bootstrap standard error estimates (SE), initial estimates for the estimation algorithm (Init), convergence criterion (ConvergeLoss), latent class membership probability (ClassProb), predicted number of recurrent events (PostPredict), relative entropy of the fitted model (Entropy), p-value (pvalue) for SLCARE.
}
\seealso{
\code{\link[=SLCARE]{SLCARE()}}
}
