% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLCARE-package.R
\docType{package}
\name{SLCARE-package}
\alias{SLCARE-package}
\alias{_PACKAGE}
\title{SLCARE: Semiparametric Latent Class Analysis of Recurrent Events}
\description{
Efficient R package for latent class analysis of recurrent events, based on the semiparametric multiplicative intensity model by Zhao et al. (2022) \doi{10.1111/rssb.12499}. SLCARE returns estimates for non-functional model parameters along with the associated variance estimates and p-values. Visualization tools are provided to depict the estimated functional model parameters and related functional quantities of interest. SLCARE also delivers a model checking plot to help assess the adequacy of the fitted model.
}
\author{
\strong{Maintainer}: Qi Yu \email{qi.yu2@emory.edu}

Authors:
\itemize{
  \item Limin Peng \email{lpeng@emory.edu}
}

}
\keyword{internal}
