% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLCARE_simdat.R
\docType{data}
\name{SLCARE_simdat}
\alias{SLCARE_simdat}
\title{Simulated dataset}
\format{
A data frame with 478 rows and 5 variables.
}
\usage{
data(SLCARE_simdat)
}
\description{
A dataset simulated from a real world dataset
\describe{
  \item{id}{subjects identification}
  \item{time}{time recored including event and longest followup time (censoring)}
  \item{event}{recurrent event indicator; 1 if a recurrent event is recorded}
  \item{x1}{a dummy baseline covariate}
  \item{x2}{a continuous baseline covariate range from 0 to 1}
}
}
