% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SLCARE}
\alias{summary.SLCARE}
\title{Summary Results for SLCARE}
\usage{
\method{summary}{SLCARE}(object, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{SLCARE}.}

\item{digits}{minimal number of significant digits.}

\item{...}{other arguments.}
}
\description{
Summary results for \code{SLCARE} object including regression coefficients, corresponding standard error estimates and relative entropy of the fitted model.
}
\seealso{
\code{\link[=SLCARE]{SLCARE()}}
}
