% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha_weightedregression.R
\name{update_alpha}
\alias{update_alpha}
\title{Estimation algorithm - updating alpha}
\usage{
update_alpha(alpha, beta, d, Z, mu_censor, gamma = 0)
}
\arguments{
\item{alpha}{a matrix of alpha before updating - regression coefficient for multinomial logistic regression model}

\item{beta}{a matrix of beta before updating - class specific parameters for recurrent model}

\item{d}{a vector of observed recurrent events for subjects of interest}

\item{Z}{a vector of time-independent corvariates}

\item{mu_censor}{a vector of estimated mu(C), where C is a vector of censoring time}

\item{gamma}{individual frailty. 0 represents the frailty equals 1 and k represents the frailty follows gamma(k,k)}
}
\value{
a matrix of updated alpha - regression coefficient for multinomial logistic regression model
}
\description{
Updating alpha in estimation procedure. Updating alpha by fitting a weighted multinomial regression.
}
